void ler inteiro adicionada
